test: 질문 조회 API 테스트 코드 수정
