+++ application/src/test/java/oxahex/asker/user/UserControllerTest.java
@@ -60,81 +60,6 @@ class UserControllerTest extends MockUser {
     this.answerUser = userRepository.save(newUser("answerer", "1234567890"));
   }
 
-  @Test
-  @WithUserDetails(value = "answerer@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
-  @DisplayName("질문 내역 조회 API 접근 성공 - 로그인 유저의 경우 질문 목록 조회 API 접근이 가능하다.")
-  public void getReceivedAsks_success_with_login_user() throws Exception {
-
-    // given: 받은 질문 없음
-    given(userService.getReceivedAsks(anyLong()))
-        .willReturn(new UserAsksDto());
-
-    // when
-    String url = "/api/users/" + answerUser.getId() + "/asks";
-    ResultActions resultActions = mockMvc.perform(get(url)
-        .contentType(MediaType.APPLICATION_JSON));
-
-    // then
-    resultActions.andExpect(status().isOk());
-  }
-
-  @Test
-  @DisplayName("질문 내역 조회 API 접근 실패 - 비로그인 유저의 경우 질문 목록 조회 API 접근이 불가하다.")
-  public void getReceivedAsks_failure_without_login() throws Exception {
-
-    // given: 받은 질문 없음
-    given(userService.getReceivedAsks(anyLong()))
-        .willReturn(new UserAsksDto());
-
-    // when
-    String url = "/api/users/" + answerUser.getId() + "/asks";
-    ResultActions resultActions = mockMvc.perform(get(url)
-        .contentType(MediaType.APPLICATION_JSON));
-
-    // then
-    resultActions.andExpect(status().isUnauthorized());
-  }
-
-  @Test
-  @WithUserDetails(value = "answerer@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
-  @DisplayName("질문 내역 조회 시 받은 질문이 없는 경우 Empty List를 반환한다.")
-  public void getReceivedAsks_success_if_null() throws Exception {
-
-    // given: 받은 질문 없음
-    UserAsksDto userAsksDto = UserDto.fromEntityToUserAsks(answerUser, new ArrayList<Dispatch>());
-    given(userService.getReceivedAsks(anyLong()))
-        .willReturn(userAsksDto);
-
-    // when
-    String url = "/api/users/" + answerUser.getId() + "/asks";
-    ResultActions resultActions = mockMvc.perform(get(url)
-        .contentType(MediaType.APPLICATION_JSON));
-
-    // then
-    resultActions.andExpect(status().isOk());
-    resultActions.andExpect(jsonPath("$.data.id").value(answerUser.getId()));
-    resultActions.andExpect(jsonPath("$.data.name").value(answerUser.getName()));
-    // 질문 목록은 Empty List
-    resultActions.andExpect(jsonPath("$.data.asks").isArray());
-    resultActions.andExpect(jsonPath("$.data.asks").isEmpty());
-  }
-
-  @Test
-  @WithUserDetails(value = "answerer@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
-  @DisplayName("질문 내역 조회 실패 - 로그인 유저의 경우 본인에게 온 질문만 확인할 수 있다.")
-  public void getReceivedAsk_failure_no_authority_to_access() throws Exception {
-
-    // given: 다른 유저의 질문 목록 요청
-    long otherUserId = answerUser.getId() + 1;
-    String url = "/api/users/" + otherUserId + "/asks";
-
-    // when
-    ResultActions resultActions = mockMvc.perform(get(url)
-        .contentType(MediaType.APPLICATION_JSON));
-
-    // then
-    resultActions.andExpect(status().isUnauthorized());
-  }
 
   @Test
   @DisplayName("답변 내역 조회 API 접근 성공 - 비로그인 유저의 경우 답변 내역 조회 API에 접근할 수 있다.")
